修改数据库中type的类型为int，外键添加on delete cascade 修复抵押物管理接口bug，除toBC以外其他的应该都能正常使用

+++ src/main/java/cn/enderqiu/bcinvestrebuild/app/GuarantyManagement/entity/vo/LandVO.java
@@ -8,13 +8,13 @@ public class LandVO extends GuarantyVO{
     @ApiModelProperty(value = "地址" ,required = true)
     private String addr;
     @ApiModelProperty(value = "面积" ,required = true)
-    private int area;
+    private long area;
 
-    public void setArea(int area) {
+    public void setArea(long area) {
         this.area = area;
     }
 
-    public int getArea() {
+    public long getArea() {
         return area;
     }
